move addInstructor function to applicationController and change name AddInstructorController to InstructorController
