+++ app/Http/Controllers/Admin/ApplicationController.php
@@ -2,13 +2,15 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\BaseController;
-use App\Http\Controllers\Controller;
-use App\Http\Requests\ApplicationRequest;
-use App\Http\Resources\ApplicationResource;
+use App\Models\User;
+use App\Models\Instructor;
 use App\Models\Application;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
+use App\Http\Controllers\Controller;
+use App\Http\Controllers\BaseController;
+use App\Http\Requests\ApplicationRequest;
+use App\Http\Requests\InstructorRequest;
+use App\Http\Resources\ApplicationResource;
 
 class ApplicationController extends BaseController
 {
@@ -16,8 +18,6 @@ class ApplicationController extends BaseController
     {
         return $this->success(ApplicationResource::collection(Application::with('cv_form')->get()),'All applications');
     }
-
-
     public function store(ApplicationRequest $request)
     {
         $application = new Application();
@@ -37,4 +37,39 @@ class ApplicationController extends BaseController
         $application->delete();
         return $this->success([],'deleted',config('http_status_code.no_content'));
     }
+
+    public function createInstructorId(){
+        $instructor = Instructor::select('instructor_id')
+            ->orderByDesc('instructor_id')
+            ->value('instructor_id');
+            $instructorIdOnly = substr($instructor,2);
+        if($instructorIdOnly){
+            $instructorId = str_pad((int)$instructorIdOnly +1 ,4,"0",STR_PAD_LEFT);
+        }else{
+            $instructorId = config('instructorid.id');
+        }
+        return "I-".$instructorId;
+    }
+
+    public function removedApplication($email){
+        $application = Application::where('email',$email)->first();
+        $application->delete();
+    }
+
+    public function addInstructor(InstructorRequest $request){
+        $user = new User();
+        $user->name = $request->name;
+        $user->email = $request->email;
+        $user->password = $request->password;
+        $user->save();
+        $instructor = new Instructor();
+        $instructor->instructor_id =$this->createInstructorId();
+        $instructor->user_id = $user->id;
+        $instructor->cv_id = $request->cv_id;
+        $instructor->phone = $request->phone;
+        $instructor->address = $request->address;
+        $instructor->save();
+        $this->removedApplication($user->email);
+        return $this->success($instructor,'Created',config('http_status_code.created'));
+   }
 }
