+++ routes/admin.php
@@ -9,7 +9,7 @@ use App\Http\Controllers\Admin\CareerController;
 use App\Http\Controllers\Admin\CourseController;
 use App\Http\Controllers\Admin\CvFormController;
 use App\Http\Controllers\Admin\ApplicationController;
-use App\Http\Controllers\Admin\AddInstructorController;
+use App\Http\Controllers\Admin\InstructorController;
 
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     $user = $request->user();
@@ -18,12 +18,10 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::apiResource('courses',CourseController::class);
 Route::apiResource('applications',ApplicationController::class);
-Route::post('applications',[ApplicationController::class,'store']);
 Route::apiResource('careers',CareerController::class);
-Route::apiResource('applicatons',ApplicationController::class);
 Route::apiResource('cv-forms',CvFormController::class);
 Route::post('images',[ImageController::class,'store']);
 Route::post('files',[FileController::class,'store']);
-Route::post('instructors',[AddInstructorController::class,'store']);
-Route::apiResource('instructors',AddInstructorController::class);
+Route::post('applications/add-instructor',[ApplicationController::class,'addInstructor']);
+Route::apiResource('instructors',InstructorController::class);
 ?>
